+++ src/App.vue
@@ -12,7 +12,7 @@ export default {
   components: {NavigationBar},
   data() {
     return {
-      searchQuery: 'Singen'
+      searchQuery: ''
     }
   },
   provide() {
@@ -32,6 +32,17 @@ export default {
   color: #2c3e50;
   margin-top: 60px;
   padding: 1em;
+}
+
+.noselect {
+  -webkit-touch-callout: none; /* iOS Safari */
+  -webkit-user-select: none; /* Safari */
+  -khtml-user-select: none; /* Konqueror HTML */
+  -moz-user-select: none; /* Old versions of Firefox */
+  -ms-user-select: none; /* Internet Explorer/Edge */
+  user-select: none;
+  /* Non-prefixed version, currently
+                                   supported by Chrome, Edge, Opera and Firefox */
 }
 
 :root {