+++ src/mountEntrypoint.js
@@ -1,6 +1,7 @@
 import app from './createEntrypoint'
 import './store/createVuex'
 import './router/createRouter'
+import '@/assets/font_awesome'
 
 import search from '@/components/common/search.vue'